Extract out a common cache filesystem to share between on-disk and in-memory
