+++ src/disk_cache_filesystem.cpp
@@ -90,6 +90,8 @@ string GetLocalCacheFile(const string &cache_directory,
 
 // Attempt to cache [chunk] to local filesystem, if there's sufficient disk
 // space available.
+//
+// TODO(hjiang): Document local cache file pattern and its eviction policy.
 void CacheLocal(const CacheReadChunk &chunk, FileSystem &local_filesystem,
                 const FileHandle &handle, const string &cache_directory,
                 const string &local_cache_file) {
@@ -130,36 +132,13 @@ void CacheLocal(const CacheReadChunk &chunk, FileSystem &local_filesystem,
 
 } // namespace
 
-DiskCacheFileHandle::DiskCacheFileHandle(
-    unique_ptr<FileHandle> internal_file_handle_p, DiskCacheFileSystem &fs)
-    : FileHandle(fs, internal_file_handle_p->GetPath(),
-                 internal_file_handle_p->GetFlags()),
-      internal_file_handle(std::move(internal_file_handle_p)) {}
-
 DiskCacheFileSystem::DiskCacheFileSystem(
     unique_ptr<FileSystem> internal_filesystem_p,
     OnDiskCacheConfig cache_config_p)
-    : cache_config(std::move(cache_config_p)),
-      local_filesystem(FileSystem::CreateLocal()),
-      internal_filesystem(std::move(internal_filesystem_p)) {
+    : CacheFileSystem(std::move(internal_filesystem_p)),
+      cache_config(std::move(cache_config_p)),
+      local_filesystem(FileSystem::CreateLocal()) {
   local_filesystem->CreateDirectory(cache_config.on_disk_cache_directory);
-}
-
-void DiskCacheFileSystem::Read(FileHandle &handle, void *buffer,
-                               int64_t nr_bytes, idx_t location) {
-  ReadImpl(handle, buffer, nr_bytes, location, DEFAULT_BLOCK_SIZE);
-}
-int64_t DiskCacheFileSystem::Read(FileHandle &handle, void *buffer,
-                                  int64_t nr_bytes) {
-  const int64_t bytes_read = ReadImpl(
-      handle, buffer, nr_bytes, handle.SeekPosition(), DEFAULT_BLOCK_SIZE);
-  handle.Seek(handle.SeekPosition() + bytes_read);
-  return bytes_read;
-}
-int64_t DiskCacheFileSystem::ReadForTesting(FileHandle &handle, void *buffer,
-                                            int64_t nr_bytes, idx_t location,
-                                            uint64_t block_size) {
-  return ReadImpl(handle, buffer, nr_bytes, location, block_size);
 }
 
 void DiskCacheFileSystem::ReadAndCache(FileHandle &handle, char *buffer,
@@ -277,7 +256,7 @@ void DiskCacheFileSystem::ReadAndCache(FileHandle &handle, char *buffer,
         cache_read_chunk.content =
             CreateResizeUninitializedString(cache_read_chunk.chunk_size);
       }
-      auto &disk_cache_handle = handle.Cast<DiskCacheFileHandle>();
+      auto &disk_cache_handle = handle.Cast<CacheFileSystemHandle>();
       internal_filesystem->Read(
           *disk_cache_handle.internal_file_handle,
           const_cast<char *>(cache_read_chunk.content.data()),
@@ -296,24 +275,6 @@ void DiskCacheFileSystem::ReadAndCache(FileHandle &handle, char *buffer,
     D_ASSERT(cur_thd.joinable());
     cur_thd.join();
   }
-}
-
-int64_t DiskCacheFileSystem::ReadImpl(FileHandle &handle, void *buffer,
-                                      int64_t nr_bytes, idx_t location,
-                                      uint64_t block_size) {
-  const auto file_size = handle.GetFileSize();
-
-  // No more bytes to read.
-  if (location == file_size) {
-    return 0;
-  }
-
-  const int64_t bytes_to_read =
-      std::min<int64_t>(nr_bytes, file_size - location);
-  ReadAndCache(handle, static_cast<char *>(buffer), location, bytes_to_read,
-               file_size, block_size);
-
-  return bytes_to_read;
 }
 
 } // namespace duckdb
